requirements.txt
fastapi
uvicorn
onnxruntime
numpy
pillow
python-multipart
onnx 
google-cloud-storage
google-cloud-firestore